move 'either' tests to separate file
