+++ types/ramda/ramda-tests.ts
@@ -2140,21 +2140,6 @@ class Rectangle {
 });
 
 () => {
-    function gt10(x: number) {
-        return x > 10;
-    }
-
-    function even(x: number) {
-        return x % 2 === 0;
-    }
-
-    const f = R.either(gt10, even);
-    const g = R.either(gt10)(even);
-    f(101); // => true
-    f(8); // => true
-};
-
-() => {
     // Flatten all arrays in the list but leave other values alone.
     const flattenArrays = R.map(R.ifElse(Array.isArray, R.flatten, R.identity));
